sdk/reader: split claude tool-result event extraction into a submodule
Extract the tool_result_event cluster (tool_result-block events, system
subagent-notification events, payload measurement/hashing/truncation
detection, and tool-replacement metadata recovery) out of claude.rs into
claude/tool_results.rs. Pure move; the ReplacementMeta/Measured structs and
their fields are pub(super) so the parse engine and tests still reach them.
first_present joins string_field as a pub(super) accessor. Test-only
helpers (measure_tool_result, detect_truncation_marker) are imported in the
test module so the lib build stays warning-clean.
